updates for readme
